+++ packages/spacecat-shared-data-access/src/models/audit/audit.model.js
@@ -31,6 +31,7 @@ class Audit extends BaseModel {
     404: '404',
     SITEMAP: 'sitemap',
     CANONICAL: 'canonical',
+    REDIRECT_CHAINS: 'redirect-chains',
     BROKEN_BACKLINKS: 'broken-backlinks',
     BROKEN_INTERNAL_LINKS: 'broken-internal-links',
     EXPERIMENTATION: 'experimentation',
@@ -41,6 +42,7 @@ class Audit extends BaseModel {
     EXPERIMENTATION_ESS_MONTHLY: 'experimentation-ess-monthly',
     EXPERIMENTATION_OPPORTUNITIES: 'experimentation-opportunities',
     META_TAGS: 'meta-tags',
+    LLM_ERROR_PAGES: 'llm-error-pages',
     COSTS: 'costs',
     STRUCTURED_DATA: 'structured-data',
     STRUCTURED_DATA_AUTO_SUGGEST: 'structured-data-auto-suggest',
@@ -51,8 +53,10 @@ class Audit extends BaseModel {
     SECURITY_CSP: 'security-csp',
     PAID: 'paid',
     HREFLANG: 'hreflang',
+    HEADINGS: 'headings',
     PAID_TRAFFIC_ANALYSIS_WEEKLY: 'paid-traffic-analysis-weekly',
     PAID_TRAFFIC_ANALYSIS_MONTHLY: 'paid-traffic-analysis-monthly',
+    READABILITY: 'readability',
   };
 
   static AUDIT_TYPE_PROPERTIES = {
@@ -73,6 +77,7 @@ class Audit extends BaseModel {
   static AUDIT_STEP_DESTINATIONS = {
     CONTENT_SCRAPER: 'content-scraper',
     IMPORT_WORKER: 'import-worker',
+    SCRAPE_CLIENT: 'scrape-client',
   };
 
   /**
@@ -86,8 +91,10 @@ class Audit extends BaseModel {
    *   [Audit.AUDIT_STEP_DESTINATIONS.IMPORT_WORKER]: {
    *     getQueueUrl: function,
    *     formatPayload: function
-   *   }
-   * }}
+   *   },
+   *   [Audit.AUDIT_STEP_DESTINATIONS.SCRAPE_CLIENT]: {
+   *   formatPayload: function
+   * }}}
    */
   static AUDIT_STEP_DESTINATION_CONFIGS = {
     [Audit.AUDIT_STEP_DESTINATIONS.IMPORT_WORKER]: {
@@ -151,6 +158,26 @@ class Audit extends BaseModel {
         options: stepResult.options || {},
         completionQueueUrl: stepResult.completionQueueUrl || context.env?.AUDIT_JOBS_QUEUE_URL,
         auditContext,
+      }),
+    },
+    [Audit.AUDIT_STEP_DESTINATIONS.SCRAPE_CLIENT]: {
+      /**
+       *
+       * @param stepResult - The result of the audit step.
+       * @param auditContext - The audit context.
+       * @param context - The context object.
+       * @returns {object} - The formatted payload for the scrape client.
+       */
+      formatPayload: (stepResult, auditContext, context) => ({
+        urls: stepResult.urls.map((urlObj) => urlObj.url),
+        processingType: stepResult.processingType || 'default',
+        options: stepResult.options || {},
+        maxScrapeAge: stepResult.maxScrapeAge || 24,
+        auditData: {
+          siteId: stepResult.siteId,
+          completionQueueUrl: stepResult.completionQueueUrl || context.env?.AUDIT_JOBS_QUEUE_URL,
+          auditContext,
+        },
       }),
     },
   };